"Reorganizando los modulos: asignatura, facultad, cliente y programa"
