+++ src/app/modulos/modulo-asignatura/asignatura-popup/asignatura-popup.component.ts
@@ -13,35 +13,38 @@ import {
   Validators,
 } from '@angular/forms';
 import { ToastService } from 'src/app/service/toast.service';
-import { ProgramaService } from 'src/app/service/programa.service';
 import { Facultad } from 'src/app/Model/Facultad';
-import { FacultadService } from 'src/app/service/facultad.service';
+import { AsignaturaService } from 'src/app/modulos/modulo-asignatura/service/asignatura.service';
+import { Asignatura } from 'src/app/Model/Asignatura';
+import { ProgramaService } from '../../modulo-programa/service/ProgramaService';
 
 @Component({
-  selector: 'app-programa-popup',
-  templateUrl: './programa-popup.component.html',
-  styleUrls: ['./programa-popup.component.css'],
+  selector: 'app-asignatura-popup',
+  templateUrl: './asignatura-popup.component.html',
+  styleUrls: ['./asignatura-popup.component.css']
 })
-export class ProgramaPopupComponent implements OnInit {
+export class AsignaturaPopupComponent implements OnInit {
   submit = false;
   hide = true;
   lista_facultad!: Facultad[];
+  lista_asignatura!: Asignatura[];
   inputdata: any;
-  idPrograma: any = null;
-  dataPrograma!: Programa;
+  idAsignatura: any = null;
+  dataAsignatura!: Asignatura;
   editdata: any = null;
   closemessage = 'closed using directive';
 
   constructor(
     @Inject(MAT_DIALOG_DATA) public data: any,
-    private ref: MatDialogRef<ProgramaPopupComponent>,
+    private ref: MatDialogRef<AsignaturaPopupComponent>,
     private formBuilder: FormBuilder,
     private service: ProgramaService,
-    private _servicioFacultad: FacultadService,
+    private _servicioAsignatura: AsignaturaService,
     private _servicioToast: ToastService
   ) {
-    this.cargar_facultades();
+    this.cargar_asignatura();
   }
+
 
   ngOnInit(): void {
     this.inputdata = this.data;
@@ -50,21 +53,21 @@ export class ProgramaPopupComponent implements OnInit {
     }
   }
 
-  cargar_facultades() {
-    this._servicioFacultad.GetFacultadActivo().subscribe((res) => {
-      this.lista_facultad = res;
+  cargar_asignatura() {
+    this._servicioAsignatura.obtenerAsignaturaActivo().subscribe((res) => {
+      this.lista_asignatura = res;
     });
   }
 
   setpopupdata(code: any) {
-    this.service.obtenerProgramaByCode(code).subscribe((item) => {
+    this._servicioAsignatura.obtenerAsignaturaByCode(code).subscribe((item) => {
       console.log(item);
-      this.idPrograma = code;
+      this.idAsignatura = code;
       this.editdata = item;
       this.myform.setValue({
         nombre: this.editdata.nombre,
         descripcion: this.editdata.descripcion,
-        idfacultad: this.editdata.idfacultad,
+        codigo: this.editdata.codigo,
       });
     });
   }
@@ -76,24 +79,24 @@ export class ProgramaPopupComponent implements OnInit {
   myform = new FormGroup({
     nombre: new FormControl('', [Validators.required]),
     descripcion: new FormControl('', [Validators.required]),
-    idfacultad: new FormControl('', [Validators.required]),
+    codigo: new FormControl('', [Validators.required]),
   });
 
   get f(): { [key: string]: AbstractControl } {
     return this.myform.controls;
   }
 
-  guardarPrograma() {
+  guardarAsignatura() {
     this.submit = true;
     if (this.myform.valid){
       if (this.editdata != null) {
-        this.dataPrograma = {
-          id: this.idPrograma,
+        this.dataAsignatura = {
+          id: this.idAsignatura,
           nombre: this.myform.value.nombre!,
           descripcion: this.myform.value.descripcion!,
-          idfacultad: this.myform.value.idfacultad!,
+          codigo: this.myform.value.codigo!,
         };
-        this.service.actualizarPrograma(this.dataPrograma).subscribe({
+        this._servicioAsignatura.actualizarAsignatura(this.dataAsignatura).subscribe({
           next: (response) => {
             this._servicioToast.mostrarExito(
               'Actualizado correctamente.',
@@ -113,7 +116,7 @@ export class ProgramaPopupComponent implements OnInit {
           },
         });
       } else {
-        this.service.guardarPrograma(this.myform.value).subscribe({
+        this._servicioAsignatura.guardarAsignatura(this.myform.value).subscribe({
           next: (response) => {
             this._servicioToast.mostrarExito(
               'Registrado correctamente.',
@@ -140,12 +143,14 @@ export class ProgramaPopupComponent implements OnInit {
     this.submit = false;
     this.myform.reset();
     this.editdata = null;
-    this.idPrograma = null;
-    this.dataPrograma = {
-      id: 0,
+    this.idAsignatura = null;
+    this.dataAsignatura = {
+      id: "0",
       nombre: '',
       descripcion: '',
-      idfacultad: '',
+      codigo: '',
     };
   }
+
+
 }
