+++ src/app/modulos/modulo-facultad/popup/facultad-popup.component.css
