--- src/app/service/asignatura.service.spec.ts
@@ -1,16 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { AsignaturaService } from './asignatura.service';
-
-describe('AsignaturaService', () => {
-  let service: AsignaturaService;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({});
-    service = TestBed.inject(AsignaturaService);
-  });
-
-  it('should be created', () => {
-    expect(service).toBeTruthy();
-  });
-});
