--- src/app/service/perfil.service.spec.ts
@@ -1,16 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { PerfilService } from './perfil.service';
-
-describe('PerfilService', () => {
-  let service: PerfilService;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({});
-    service = TestBed.inject(PerfilService);
-  });
-
-  it('should be created', () => {
-    expect(service).toBeTruthy();
-  });
-});
